pyproject.toml
[tool.poetry]
name = "livraria-app"
version = "0.1.0"
description = ""
authors = ["Felipe Sampaio <felipe.sampaio@farroupilha.ifrs.edu.br>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
